Log errors of WEL not clear
Existing code makes the assumption that if the write enable bit has not
cleared after a write operation is finished, it must be because the
command was not supported and never did anything.

This assumption is wrong.

It is observed that flash will become busy, perform the write operation
successfully, clear the busy bit in the status register, and then
non-atomically clear the write enable bit shortly afterward.

Thus reading the status in this window before WEL is cleared results in
a status of non-busy but WEL still set.  This is incorrectly being
returned as an error.