Set flash mode to requested value
If the flash was in SPI mode and requested to change to QIO mode, the
flash configuration would be changed.

But if SPI mode was requested and the chip was in QIO mode, it would not
be changed.

The QE bit is non-volatile.  So resetting flash will not change it back
to non-quad.

The result is that after putting the device into quad mode once, it
would never work correctly in SPI mode again.