Support 32 bit addressing in MXIC driver
Each SPI driver is different, despite being nearly identical with only a
couple easily parameterized differences.

So while the winbond driver already supports 32 bit address, the MXIC
driver doesn't.  So we cut&paste the code into here.  But it has
problems that need to be fixed.  And a bunch of junk in the read
function with unneeded extra buffers and alignment is removed for a nice
speedup.

But none of the other drivers benefit, because they are separate.