Don't fail when WEL is still set after write command
This bit is supposed to automatically clear after a successfull write
command.

However, this appears to not be atomic with the busy bit becoming unset.
Maybe this is a bug in the flash chip?

This means the check done in wait_idle for unsupported write commands
returns false positives.

Log a message, but don't fail when it happens.  Since the flash chip
supports the commands we use, and does not randomly decide to unsupport
them, we don't need to worry about real postives.